feat(server/model): rewrite Question model
