--- server/app/model/MultipleChoiceQuestion.js
@@ -1,22 +1,0 @@
-const mongoose = require('mongoose')
-
-const multipleChoiceQuestionSchema = new mongoose.Schema({
-  title: {
-    type: String,
-    required: true,
-  },
-  subtitle: String,
-  options: {
-    type: [String],
-    required: true,
-  },
-  answer: {
-    type: Number,
-    required: true,
-  },
-})
-
-
-const MultipleChoiceQuestion = mongoose.model('MultipleChoiceQuestion', multipleChoiceQuestionSchema)
-
-module.exports = MultipleChoiceQuestion
