+++ server/app/utils/constants.js
@@ -1,5 +1,5 @@
-const MULTIPLE_CHOICE_QUESTION = 0
-const MULTIPLE_SELECTION_QUESTION = 1
+const MULTIPLE_CHOICE_QUESTION = 'MultipleChoiceQuestion'
+const MULTIPLE_SELECTION_QUESTION = 'MultipleSelectionQuestion'
 
 module.exports = {
   MULTIPLE_CHOICE_QUESTION,
